devops/release-management/tools/release-manager/requirements.txt
boto3==1.34.11
packaging==23.2
PyGithub==2.3.0
pypi-json==0.4.0
requests==2.32.2
semver==3.0.2
tabulate==0.9.0
termcolor==2.3.0
